+++ lib/recurly/client.php
@@ -3325,6 +3325,45 @@ endpoint to obtain only the newly generated `UniqueCouponCodes`.
     }
   
     /**
+     * List the external payment phases on an external subscription
+     *
+     * @param string $external_subscription_id External subscription id
+     * @param array  $options                  Associative array of optional parameters
+     *
+     * Supported optional query string parameters:
+     *
+     * - $options['params']['sort'] (string): Sort field. You *really* only want to sort by `updated_at` in ascending
+     *        order. In descending order updated records will move behind the cursor and could
+     *        prevent some records from being returned.
+     * - $options['params']['limit'] (int): Limit number of records 1-200.
+     * - $options['params']['order'] (string): Sort order.
+     *
+     * @return \Recurly\Pager A list of the the external_payment_phases on a site.
+     * @link   https://developers.recurly.com/api/v2021-02-25#operation/list_external_subscription_external_payment_phases
+     */
+    public function listExternalSubscriptionExternalPaymentPhases(string $external_subscription_id, array $options = []): \Recurly\Pager
+    {
+        $path = $this->interpolatePath("/external_subscriptions/{external_subscription_id}/external_payment_phases", ['external_subscription_id' => $external_subscription_id]);
+        return new \Recurly\Pager($this, $path, $options);
+    }
+  
+    /**
+     * Fetch an external payment_phase
+     *
+     * @param string $external_subscription_id  External subscription id
+     * @param string $external_payment_phase_id External payment phase ID, e.g. `a34ypb2ef9w1`.
+     * @param array  $options                   Associative array of optional parameters
+     *
+     * @return \Recurly\Resources\ExternalPaymentPhase Details for an external payment_phase.
+     * @link   https://developers.recurly.com/api/v2021-02-25#operation/get_external_subscription_external_payment_phase
+     */
+    public function getExternalSubscriptionExternalPaymentPhase(string $external_subscription_id, string $external_payment_phase_id, array $options = []): \Recurly\Resources\ExternalPaymentPhase
+    {
+        $path = $this->interpolatePath("/external_subscriptions/{external_subscription_id}/external_payment_phases/{external_payment_phase_id}", ['external_subscription_id' => $external_subscription_id, 'external_payment_phase_id' => $external_payment_phase_id]);
+        return $this->makeRequest('GET', $path, [], $options);
+    }
+  
+    /**
      * List entitlements granted to an account
      *
      * @param string $account_id Account ID or code. For ID no prefix is used e.g. `e28zov4fw0v2`. For code use prefix `code-`, e.g. `code-bob`.
@@ -3364,6 +3403,35 @@ endpoint to obtain only the newly generated `UniqueCouponCodes`.
     public function listAccountExternalSubscriptions(string $account_id, array $options = []): \Recurly\Pager
     {
         $path = $this->interpolatePath("/accounts/{account_id}/external_subscriptions", ['account_id' => $account_id]);
+        return new \Recurly\Pager($this, $path, $options);
+    }
+  
+    /**
+     * Fetch a business entity
+     *
+     * @param string $business_entity_id Business Entity ID. For ID no prefix is used e.g. `e28zov4fw0v2`. For code use prefix `code-`, e.g. `code-entity1`.
+     * @param array  $options            Associative array of optional parameters
+     *
+     * @return \Recurly\Resources\BusinessEntity Business entity details
+     * @link   https://developers.recurly.com/api/v2021-02-25#operation/get_business_entity
+     */
+    public function getBusinessEntity(string $business_entity_id, array $options = []): \Recurly\Resources\BusinessEntity
+    {
+        $path = $this->interpolatePath("/business_entities/{business_entity_id}", ['business_entity_id' => $business_entity_id]);
+        return $this->makeRequest('GET', $path, [], $options);
+    }
+  
+    /**
+     * List business entities
+     *
+     * @param array $options Associative array of optional parameters
+     *
+     * @return \Recurly\Pager List of all business entities on your site.
+     * @link   https://developers.recurly.com/api/v2021-02-25#operation/list_business_entities
+     */
+    public function listBusinessEntities(array $options = []): \Recurly\Pager
+    {
+        $path = $this->interpolatePath("/business_entities", []);
         return new \Recurly\Pager($this, $path, $options);
     }
   
@@ -3440,6 +3508,49 @@ endpoint to obtain only the newly generated `UniqueCouponCodes`.
     {
         $path = $this->interpolatePath("/gift_cards/{redemption_code}/redeem", ['redemption_code' => $redemption_code]);
         return $this->makeRequest('POST', $path, $body, $options);
+    }
+  
+    /**
+     * List a business entity's invoices
+     *
+     * @param string $business_entity_id Business Entity ID. For ID no prefix is used e.g. `e28zov4fw0v2`. For code use prefix `code-`, e.g. `code-entity1`.
+     * @param array  $options            Associative array of optional parameters
+     *
+     * Supported optional query string parameters:
+     *
+     * - $options['params']['ids'] (array): Filter results by their IDs. Up to 200 IDs can be passed at once using
+     *        commas as separators, e.g. `ids=h1at4d57xlmy,gyqgg0d3v9n1,jrsm5b4yefg6`.
+     *        
+     *        **Important notes:**
+     *        
+     *        * The `ids` parameter cannot be used with any other ordering or filtering
+     *          parameters (`limit`, `order`, `sort`, `begin_time`, `end_time`, etc)
+     *        * Invalid or unknown IDs will be ignored, so you should check that the
+     *          results correspond to your request.
+     *        * Records are returned in an arbitrary order. Since results are all
+     *          returned at once you can sort the records yourself.
+     * - $options['params']['limit'] (int): Limit number of records 1-200.
+     * - $options['params']['order'] (string): Sort order.
+     * - $options['params']['sort'] (string): Sort field. You *really* only want to sort by `updated_at` in ascending
+     *        order. In descending order updated records will move behind the cursor and could
+     *        prevent some records from being returned.
+     * - $options['params']['begin_time'] (string): Inclusively filter by begin_time when `sort=created_at` or `sort=updated_at`.
+     *        **Note:** this value is an ISO8601 timestamp. A partial timestamp that does not include a time zone will default to UTC.
+     * - $options['params']['end_time'] (string): Inclusively filter by end_time when `sort=created_at` or `sort=updated_at`.
+     *        **Note:** this value is an ISO8601 timestamp. A partial timestamp that does not include a time zone will default to UTC.
+     * - $options['params']['type'] (string): Filter by type when:
+     *        - `type=charge`, only charge invoices will be returned.
+     *        - `type=credit`, only credit invoices will be returned.
+     *        - `type=non-legacy`, only charge and credit invoices will be returned.
+     *        - `type=legacy`, only legacy invoices will be returned.
+     *
+     * @return \Recurly\Pager A list of the business entity's invoices.
+     * @link   https://developers.recurly.com/api/v2021-02-25#operation/list_business_entity_invoices
+     */
+    public function listBusinessEntityInvoices(string $business_entity_id, array $options = []): \Recurly\Pager
+    {
+        $path = $this->interpolatePath("/business_entities/{business_entity_id}/invoices", ['business_entity_id' => $business_entity_id]);
+        return new \Recurly\Pager($this, $path, $options);
     }
   
 }