+++ lib/recurly/resources/subscription_ramp_interval_response.php
@@ -12,14 +12,39 @@ use Recurly\RecurlyResource;
 // phpcs:disable
 class SubscriptionRampIntervalResponse extends RecurlyResource
 {
+    private $_ending_on;
     private $_remaining_billing_cycles;
     private $_starting_billing_cycle;
+    private $_starting_on;
     private $_unit_amount;
 
     protected static $array_hints = [
     ];
 
     
+    /**
+    * Getter method for the ending_on attribute.
+    * Date the ramp interval ends
+    *
+    * @return ?string
+    */
+    public function getEndingOn(): ?string
+    {
+        return $this->_ending_on;
+    }
+
+    /**
+    * Setter method for the ending_on attribute.
+    *
+    * @param string $ending_on
+    *
+    * @return void
+    */
+    public function setEndingOn(string $ending_on): void
+    {
+        $this->_ending_on = $ending_on;
+    }
+
     /**
     * Getter method for the remaining_billing_cycles attribute.
     * Represents how many billing cycles are left in a ramp interval.
@@ -64,6 +89,29 @@ class SubscriptionRampIntervalResponse extends RecurlyResource
     public function setStartingBillingCycle(int $starting_billing_cycle): void
     {
         $this->_starting_billing_cycle = $starting_billing_cycle;
+    }
+
+    /**
+    * Getter method for the starting_on attribute.
+    * Date the ramp interval starts
+    *
+    * @return ?string
+    */
+    public function getStartingOn(): ?string
+    {
+        return $this->_starting_on;
+    }
+
+    /**
+    * Setter method for the starting_on attribute.
+    *
+    * @param string $starting_on
+    *
+    * @return void
+    */
+    public function setStartingOn(string $starting_on): void
+    {
+        $this->_starting_on = $starting_on;
     }
 
     /**